refactor(server): split runs.preview.test.ts into focused preview stage files (warren-d951)
